+++ changelog.rst
@@ -2,6 +2,7 @@ Upcoming
 ========
 
 * Added `build-system` section to `pyproject.toml` to use the modern setuptools backend.
+* Fixed test failures when locale other than `en_US` is used.
 
 2.1.0 (2023-03-31)
 =========